[Modify] 마이페이지 대시보드 수정
- 아직 수정 완성은 아님
- 찜한 식자재 이름은 보임
- 오후에 식자재 물가 추이 그래프 수정 예정

+++ FocusOnMeal/src/main/java/com/fom/boot/app/mypage/dto/FavoriteIngredientSummaryDTO.java
@@ -19,6 +19,6 @@ public class FavoriteIngredientSummaryDTO {
     
     private String ingredientName;
     private String category;
-    private BigDecimal currentPrice; 	// 최신 가격 (int로 변경)
+    private Integer currentPrice;
     private String standardUnit;
 }
